add a quick benchmark script
This could be extended a good deal. The main thing I want to add is the ability to generate data for a couple of days at a time.

Note, it uses a modified version of the client with the module name changed to sixpack_client, b/c otherwise it conflicts with the server module.