Discard conversions from excluded clients when traffic_fraction < 1.
When traffic_fraction is < 1, some clients get the control alternative.
The participations of these excluded clients are not recorded to redis.
When there is a conversion request for an excluded client, the conversion
is not discarded and recorded to redis. When there are a couple of these
conversions by excluded clients, the number of completed conversions
becomes bigger than the number of participants, which should never be
possible. The computation of the confidence_interval relies on this
assumption and fails when the completed_count becomes bigger than
participant_count.

The solution is to discard the conversions of excluded clients as well.