Ensure atomic experiment creation.
The transaction wrapping the creation of experiment keys in
Experiment.save did not watch any of those keys.  Consequently
parallel writers could cause duplicate writes against an
experiment's mutable values, such as its alternatives list.

This adds a watch guard to the experiment's key. Note that
Experiment.is_new_record now checks the existence of the
experiment's key rather than its membership in the experiments
set; this is so that the code watches a value unique to the
experiment and not one shared by all.

On conflict the experiment's traffic fraction is persisted.